+++ extensions/defbro/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Defbro Changelog
 
+## [Added frecency sorting to browser list] - {PR_MERGE_DATE}
+
+- Uses the built-in [useFrecencySorting hook](https://developers.raycast.com/utilities/react-hooks/usefrecencysorting) to sort browsers for convenience.
+
 ## [Initial Version] - 2024-11-19